feat(appbar) move top toolbar to app tool bar
