+++ client/src/components/App/AppToolbar/AppToolbarStyles.ts
@@ -1,12 +1,25 @@
-import { makeStyles } from '@material-ui/styles';
+import { Theme } from '@material-ui/core';
+import { makeStyles } from '@material-ui/styles'
 
-import theme from 'styles/theme';
-
-const useStyles = makeStyles({
-    appToolbar: {
-        color: 'white',
-        backgroundColor: theme.palette.primary.dark
+const useStyles = makeStyles((theme: Theme) => ({
+    logoTitle: {
+        display: 'flex',
+        flexDirection: 'row',
+        flexGrow: 1
+    },
+    userSection: {
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+    },
+    greetUserText: {
+        marginLeft: theme.spacing(2),
+        flip: false
+    },
+    swalTitle: {
+        fontSize: '1.5vw',
+        fontFamily: 'Assistant',
     }
-});
+}));
 
 export default useStyles;
